+++ src/normcap/gui/system_info.py
@@ -2,40 +2,18 @@
 import functools
 import logging
 import os
-import re
-import subprocess
 import sys
 from importlib import metadata
 from pathlib import Path
-from typing import Optional
 
-from packaging import version
 from PySide6 import QtCore, QtGui, QtWidgets
 from PySide6 import __version__ as PySide6_version
 
 from normcap import __version__
 from normcap.gui.models import DesktopEnvironment, Rect, Screen
+from normcap.screengrab import gnome_shell_version
 
 logger = logging.getLogger(__name__)
-
-
-@functools.lru_cache()
-def gnome_shell_version() -> Optional[version.Version]:
-    """Get gnome-shell version (Linux, Gnome)."""
-    if sys.platform != "linux" or desktop_environment() != DesktopEnvironment.GNOME:
-        return None
-
-    shell_version = None
-    try:
-        output_raw = subprocess.check_output(["gnome-shell", "--version"], shell=False)
-        output = output_raw.decode().strip()
-        if result := re.search(r"\s+([\d.]+)", output):
-            shell_version = version.parse(result.groups()[0])
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        pass
-    except Exception as e:  # pylint: disable=broad-except
-        logger.error("Exception when trying to get gnome-shell version %s", e)
-    return shell_version
 
 
 @functools.cache